\name{MESFfast-package}
\alias{MESFfast-package}
\alias{MESFfast}
\docType{package}
\title{
    Faster Moran Eigenvector Spatial Filtering
}
\description{
    This package combines other tools and packages to more
    efficiently and quickly implement Moran Eigenvector Spatial Filtering
    (MESF) for larger data sets. MESF is a useful tool for addressing
    spatial autocorrelation.  Credit to Matthew Cooper for functions
    modified from the moranfast package and to Roger Bivand and Pedro
    Peres-Neto for functions modified from the spatialreg package.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Michelle Vered
Maintainer: Michelle Vered <me@michellevered.com>
}
\references{
     Cooper, Matthew (2025). moranfast: Calculate Moran's I quickly with low memory footprint for Big Data. https://github.com/mcooper/moranfast/

     Bivand, Roger and Pedro Peres-Neto (2008). spatialreg: ME https://github.com/r-spatial/spatialreg/blob/main/R/ME.R




}
\keyword{ package }
\seealso{
  mesf
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.
  }
}
