% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moranfast.R
\name{moranfast}
\alias{moranfast}
\title{Calculate Moran's I statistic quickly}
\usage{
moranfast(x, lat, lon, alternative = "two.sided")
}
\arguments{
\item{x}{a numeric vector of values with a spatial distribution}

\item{lat}{a numeric vector indicating the latitude for each element of x}

\item{lon}{a numeric vector indicating the longitude for each element of x}

\item{alternative}{a string specifying the alternative hypothesis to be tested against, must be one of "two.sided", "less", or "greater"}
}
\description{
Calculates Moran's I statistic, measuring spatial autocorrelation, for a
vector of values distributed across space, with locations indicated by
latitude/longitude. Utilizes improvements originally developed by Matt Cooper
to optimize the calculation for big data applications. It is memory optimized
as it calculates the distance matrix on-the-fly as needed for the I statistic
calculation, rather than defining the entire spatial weights matrix first, then
calculating the global Moran's I. Calculations are performed in C++ using Rcpp
to speed up the process over implementations calculating Moran's I that are
purely written in R.
}
